Show time effort for course items (#733)
- Show effort provided by the server for all course items
- By extension, the following change hab to be performed:
  - Remove preloading UI (shimmer) for course items
  - Split course item protocols for preloading and detailed content
  - The time effort detailed data replaces stream and text
  - Adjust/Improve storyboard file for course item cell
